Fixing of a few modules
