--- NetScoreCalculation/MetricCalculation/BusFactor.rs
@@ -1,8 +1,0 @@
- /*  Name:
-  *  Date of Last Edit:
-  *
-  *  Purpose: Calculate Bus Factor Sub Metric of a given Github Repository
-  *
-  *  Details: 
- */
-
